cartList.jsp 빈 카트 전체 주문 시 유효성 검사 추가 order에서 활용을 위해 cart 서버단 parameter 변경 orderComplete.jsp에서 새로고침 시 insert 중복되는 현상 제어 일부 필드 삭제 및 변경 order 주문 서버단 구현 1차 완료

+++ smartClosetProject/src/main/java/com/spring/client/order/vo/OrderVO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class OrderVO {
-	private int order_num = 0;				// 주문 번호
+	private long order_num = 0;				// 주문 번호
 	private int order_totalPayment = 0;		// 결제금액
 	private String order_state = "";		// 주문 상태
 	private String order_cancelReason = "";	// 주문 취소 사유
